requirements.txt
azure-functions==1.23.0 
azure-functions-durable==1.3.3
azure-core==1.35.1  
azure-ai-documentintelligence==1.0.2  
azure-search-documents==11.5.3  
openai==1.107.1  
requests==2.32.5  
beautifulsoup4==4.14.2  
python-dotenv==1.1.1 